packages/ai_service/requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-chroma
langchain-ollama
langchain-huggingface
chromadb
unstructured
python-frontmatter
python-dotenv
streamlit
